[nrf fromtree] mgmt: mcumgr: Split mgmt defines to new file
This splits defines to a separate file, which are commonly needed
by out-of-tree projects whereby the full MCUmgr include cannot be
used due to undefined types being used.

(cherry picked from commit a15752c79cba648541b87fd54d58379daa61c6db)
